Refactor RelationshipRestRepository Integration tests
Refactor the integration tests for the RelationshipRestRepository
to reduce the amount of code duplication. Many tests contained code
to create a number of variables to be used in the test, this code
was pretty much the same for every test. I refactored the integration
test class so that all variables used in multiple tests are now declared
at the top and are initialized in an @Before function. Test-specific
variables are unchanged.