Fix tests for recent item changes
The item matcher didn't match anymore because it was still looking
for the bitstream link, which isn't there anymore because the item
only holds bundles now (see commit 48ea95bc). The tests were changed
to look for a bundle link.

Other than that a few tests were also specifically looking for a
bitstream link to see if a bitstream was correctly created. These
links were changed to bundle links, since a new bitstream should
be added to a new bundle by default (if there are no bundles yet)
during creation.